requirements.txt
streamlit==1.29.0
pandas==2.1.3
openai==1.3.0
anthropic==0.7.0
chromadb==0.4.17
langchain==0.0.335
pytest==7.4.3
tqdm==4.66.1
python-dotenv==1.0.0
